Allow decorator to be unbundled.
